+++ augeias/__init__.py
@@ -19,14 +19,20 @@ def includeme(config):
                      request_method="DELETE")
     config.add_route('list_object_keys_for_container',
                      pattern='/collections/{collection_key}/containers/{container_key}', request_method="GET")
+    config.add_route('create_object_and_id', pattern='/collections/{collection_key}/containers/{container_key}',
+                     request_method="POST")
     config.add_route('update_object', pattern='/collections/{collection_key}/containers/{container_key}/{object_key}',
                      request_method="PUT")
     config.add_route('delete_object', pattern='/collections/{collection_key}/containers/{container_key}/{object_key}',
                      request_method="DELETE")
     config.add_route('get_object', pattern='/collections/{collection_key}/containers/{container_key}/{object_key}',
                      request_method="GET")
+    config.add_route('get_object_info',
+                     pattern='/collections/{collection_key}/containers/{container_key}/{object_key}/meta',
+                     request_method="GET")
 
     config.scan()
+
 
 def main(global_config, **settings):
     """ This function returns a Pyramid WSGI application.