+++ augeias/stores/PairTreeFileSystemStore.py
@@ -3,9 +3,12 @@
 This module provide a simple filesystem based store
 '''
 
-from pairtree import PairtreeStorageFactory, PartNotFoundException, ObjectNotFoundException
+from pairtree import PairtreeStorageFactory, PartNotFoundException, ObjectNotFoundException, id2path
 from augeias.stores.StoreInterface import IStore
 from augeias.stores.error import NotFoundException
+import os
+import magic
+import datetime
 
 
 class PairTreeFileSystemStore(IStore):
@@ -37,6 +40,26 @@ class PairTreeFileSystemStore(IStore):
             return container.get_bytestream(object_key)
         except PartNotFoundException:
             raise NotFoundException
+
+    def get_object_info(self, container_key, object_key):
+        '''
+        Retrieve object info (mimetype, size, time last modification) from the data store.
+
+        :param str container_key: Key of the container that the object lives in.
+        :param str object_key: Key of the object to retrieve.
+        :raises augeias.stores.error.NotFoundException: When the object or container could not be found.
+        '''
+        # todo magic.from_buffer(open(file_path).read(1048576), mime=True) cannot microsoft mimetypes
+        # https://stackoverflow.com/questions/17779560/django-python-magic-identify-ppt-docx-word-uploaded-file-as-application-zip
+        file_path = '/'.join([self.store.store_dir, 'pairtree_root', id2path(container_key), object_key])
+        if not os.path.exists(file_path):
+            raise NotFoundException
+        file_stat = os.stat(file_path)
+        return {
+            'time_last_modification': datetime.datetime.fromtimestamp(file_stat.st_mtime).isoformat(),
+            'size': file_stat.st_size,
+            'mime': magic.from_buffer(open(file_path).read(1048576), mime=True) or 'application/octet-stream'
+        }
 
     def create_object(self, container_key, object_key, object_data):
         '''