+++ augeias/views.py
@@ -24,16 +24,7 @@ def failed_validation(exc, request):
     return {'message': 'Failed validation: %s' % exc.msg}
 
 
-
 class AugeiasView(object):
-
-    @staticmethod
-    def _is_long(s):
-        try:
-            long(s)
-            return True
-        except ValueError:
-            return False
 
     def __init__(self, request):
         self.request = request
@@ -41,15 +32,6 @@ class AugeiasView(object):
     @view_config(route_name='home', renderer='json')
     def my_view(self):
         return {'project': 'augeias'}
-
-    def _get_object_data(self):
-        if 'Content-Length' not in self.request.headers or not self._is_long(self.request.headers['Content-Length']):
-            raise HTTPLengthRequired
-        content_length = long(self.request.headers['Content-Length'])
-        object_data = self.request.body_file
-        if content_length == 0:
-            raise HTTPBadRequest('body is empty')
-        return object_data
 
     @view_config(route_name='list_collections', permission='view')
     def list_collections(self):
@@ -59,15 +41,40 @@ class AugeiasView(object):
 
     @view_config(route_name='update_object', permission='edit')
     def update_object(self):
-        '''update an object in the data store'''
-        collection = self.retrieve_collection()
+        '''
+        Update an object in the data store.
+        The input object data may be:
+        - an object stream (Content-Type: application/octet-stream),
+        - or a store location (host url, collection_key, container_key and object_key) 
+            within the same Augeias instance (Content-Type: application/json). 
+        '''
+        object_data = _get_object_data(self.request)
+        collection = _retrieve_collection(self.request)
         container_key = self.request.matchdict['container_key']
         object_key = self.request.matchdict['object_key']
         if len(object_key) < 3:
             raise ValidationFailure('The object key must be 3 characters long')
-        object_data = self._get_object_data()
         collection.object_store.update_object(container_key, object_key, object_data)
         res = Response(content_type='application/json', status=200)
+        res.json_body = {
+            'container_key': container_key,
+            'object_key': object_key,
+            'uri': collection.uri_generator.generate_object_uri(
+                collection=collection.name,
+                container=container_key,
+                object=object_key)
+        }
+        return res
+
+    @view_config(route_name='create_object_and_id', permission='edit')
+    def create_object_and_id(self):
+        '''create an object in the data store and generate an id'''
+        object_data = _get_object_data(self.request)
+        collection = _retrieve_collection(self.request)
+        container_key = self.request.matchdict['container_key']
+        object_key = str(uuid.uuid4())
+        collection.object_store.update_object(container_key, object_key, object_data)
+        res = Response(content_type='application/json', status=201)
         res.json_body = {
             'container_key': container_key,
             'object_key': object_key,
@@ -81,7 +88,7 @@ class AugeiasView(object):
     @view_config(route_name='delete_object', permission='edit')
     def delete_object(self):
         '''delete an object from the data store'''
-        collection = self.retrieve_collection()
+        collection = _retrieve_collection(self.request)
         container_key = self.request.matchdict['container_key']
         object_key = self.request.matchdict['object_key']
         collection.object_store.delete_object(container_key, object_key)
@@ -99,18 +106,29 @@ class AugeiasView(object):
     @view_config(route_name='get_object', permission='view')
     def get_object(self):
         '''retrieve an object from the data store'''
-        collection = self.retrieve_collection()
+        collection = _retrieve_collection(self.request)
         container_key = self.request.matchdict['container_key']
         object_key = self.request.matchdict['object_key']
         object_data = collection.object_store.get_object(container_key, object_key)
-        res = Response(content_type='application/octet-stream', status=200)
+        content_type = collection.object_store.get_object_info(container_key, object_key)['mime']
+        res = Response(content_type=content_type, status=200)
         res.body = object_data
+        return res
+
+    @view_config(route_name='get_object_info', permission='view')
+    def get_object_info(self):
+        '''retrieve object info (mimetype, size, time last modification) from the data store'''
+        collection = _retrieve_collection(self.request)
+        container_key = self.request.matchdict['container_key']
+        object_key = self.request.matchdict['object_key']
+        res = Response(content_type='application/json', status=200)
+        res.json_body = collection.object_store.get_object_info(container_key, object_key)
         return res
 
     @view_config(route_name='list_object_keys_for_container', permission='view')
     def list_object_keys_for_container(self):
         '''list all object keys for a container in the data store'''
-        collection = self.retrieve_collection()
+        collection = _retrieve_collection(self.request)
         container_key = self.request.matchdict['container_key']
         res = Response(content_type='application/json', status=200)
         res.json_body = collection.object_store.list_object_keys_for_container(container_key)
@@ -119,7 +137,7 @@ class AugeiasView(object):
     @view_config(route_name='create_container', permission='edit')
     def create_container(self):
         '''create a new container in the data store'''
-        collection = self.retrieve_collection()
+        collection = _retrieve_collection(self.request)
         container_key = self.request.matchdict['container_key']
         collection.object_store.create_container(container_key)
         res = Response(content_type='application/json', status=200)
@@ -134,7 +152,7 @@ class AugeiasView(object):
     @view_config(route_name='create_container_and_id', permission='edit')
     def create_container_and_id(self):
         '''create a new container in the data store and generate an id'''
-        collection = self.retrieve_collection()
+        collection = _retrieve_collection(self.request)
         container_key = str(uuid.uuid4())
         collection.object_store.create_container(container_key)
         res = Response(content_type='application/json', status=201)
@@ -149,7 +167,7 @@ class AugeiasView(object):
     @view_config(route_name='delete_container', permission='edit')
     def delete_container(self):
         '''delete a container in the data store'''
-        collection = self.retrieve_collection()
+        collection = _retrieve_collection(self.request)
         container_key = self.request.matchdict['container_key']
         collection.object_store.delete_container(container_key)
         res = Response(content_type='application/json', status=200)
@@ -161,10 +179,71 @@ class AugeiasView(object):
         }
         return res
 
-    def retrieve_collection(self):
-        collection_name = self.request.matchdict['collection_key']
-        if collection_name in self.request.registry.collections:
-            collection = self.request.registry.collections[collection_name]
-        else:
-            raise HTTPNotFound('collection not found')
-        return collection
+
+# HELPERS
+
+
+def _is_long(s):
+    try:
+        long(s)
+        return True
+    except ValueError:
+        return False
+
+
+def _get_object_data(request):
+    if request.content_type == 'application/json':
+        return _get_object_data_from_json_body(request)
+    else:
+        return _get_object_data_from_stream(request)
+
+
+def _get_object_data_from_stream(request):
+    if 'Content-Length' not in request.headers or not _is_long(request.headers['Content-Length']):
+        raise HTTPLengthRequired
+    content_length = long(request.headers['Content-Length'])
+    object_data = request.body_file
+    if content_length == 0:
+        raise HTTPBadRequest('body is empty')
+    return object_data
+
+
+def _get_object_data_from_json_body(request):
+    '''
+    The json body contains the location of the input object. 
+    The body must include be the host url, collection_key, container_key and object_key.
+    '''
+    json_data = _get_json_from_request(request)
+    required_keys = ['host_url', 'collection_key', 'container_key', 'object_key']
+    missing_keys = set(required_keys) - set(json_data.keys())
+    if missing_keys:
+        raise HTTPBadRequest('\n'.join('{} is Required.'.format(key) for key in missing_keys))
+    if request.host_url != json_data['host_url']:
+        raise ValidationFailure('Host must be equal to the current host url {}.'.format(request.host))
+    collection = request.registry.collections.get(json_data['collection_key'])
+    if not collection:
+        raise HTTPBadRequest('Collection {} was not found'.format(json_data['collection_key']))
+    try:
+        object_data = collection.object_store.get_object(json_data['container_key'], json_data['object_key'])
+    except NotFoundException:
+        raise HTTPBadRequest('Container - object ({0} - {1}) combination was not found in Collection {2}'.format(
+            json_data['container_key'], json_data['object_key'], json_data['collection_key']))
+    return object_data
+
+
+def _get_json_from_request(request):
+    try:
+        return request.json_body
+    except AttributeError as e:
+        raise HTTPBadRequest(detail="Request has no json body. \n%s" % e)  # pragma: no cover
+    except ValueError as e:
+        raise HTTPBadRequest(detail="Request has incorrect json body. \n%s" % e)
+
+
+def _retrieve_collection(request):
+    collection_name = request.matchdict['collection_key']
+    if collection_name in request.registry.collections:
+        collection = request.registry.collections[collection_name]
+    else:
+        raise HTTPNotFound('collection not found')
+    return collection