+++ tests/test_functional.py
@@ -8,6 +8,7 @@ from augeias import main
 import ast
 from augeias.collections import Collection
 from augeias.stores.PairTreeFileSystemStore import PairTreeFileSystemStore
+import json
 
 here = os.path.dirname(__file__)
 settings = get_appsettings(os.path.join(here, 'conf_test.ini'))
@@ -104,6 +105,38 @@ class FunctionalTests(unittest.TestCase):
         self.assertEqual(file_size, len(res.body))
         self.assertEqual(bdata, res.body)
 
+    def test_get_object_info(self):
+        # create container and add object
+        cres = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID')
+        self.assertEqual('200 OK', cres.status)
+        self.assertIn(self.storage_location + 'collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID',
+                      cres.body)
+        testdata = os.path.join(here, '../', 'fixtures/kasteel.jpg')
+        with open(testdata, 'rb') as f:
+            bdata = f.read()
+        ores = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID/200x300', bdata)
+        self.assertIn(self.storage_location + 'collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID/200x300',
+                      ores.body)
+        self.assertEqual('200 OK', ores.status)
+
+        res = self.testapp.get('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID/200x300/meta')
+        self.assertEqual('200 OK', res.status)
+        self.assertEqual(11370, res.json_body['size'])
+        self.assertEqual('image/jpeg', res.json_body['mime'])
+        self.assertIn('time_last_modification', res.json_body)
+
+    def test_get_object_info_not_found(self):
+        # create container and add object
+        cres = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID')
+        self.assertEqual('200 OK', cres.status)
+
+        res = self.testapp.get('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID/not_found/meta',
+                               expect_errors=True)
+        self.assertEqual('404 Not Found', res.status)
+        res2 = self.testapp.get('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_NOT_FOUND/not_found/meta',
+                                expect_errors=True)
+        self.assertEqual('404 Not Found', res2.status)
+
     def test_list_object_keys_for_container(self):
         # create container and add objects
         cres = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID')
@@ -160,6 +193,77 @@ class FunctionalTests(unittest.TestCase):
         self.assertDictEqual({u'message': u'Failed validation: The object key must be 3 characters long'},
                              res.json_body)
 
+    def test_copy_object(self):
+        # create container and add object
+        cres = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID')
+        self.assertEqual('200 OK', cres.status)
+        testdata = os.path.join(here, '../', 'fixtures/kasteel.jpg')
+        with open(testdata, 'rb') as f:
+            bdata = f.read()
+        ores = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID/200x300', bdata)
+        self.assertEqual('200 OK', ores.status)
+        self.assertIn(self.storage_location + 'collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID/200x300',
+                      ores.body)
+        cres2 = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2')
+        self.assertEqual('200 OK', cres2.status)
+        json_data = json.dumps({
+            'host_url': 'http://localhost',
+            'collection_key': 'TEST_COLLECTION',
+            'container_key': 'TEST_CONTAINER_ID',
+            'object_key': '200x300'
+        })
+        ores2 = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2/kasteel',
+                                 json_data, headers={'content-type': 'application/json'})
+
+        self.assertEqual('200 OK', ores2.status)
+        res2 = self.testapp.get('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2/kasteel')
+        self.assertEqual('200 OK', res2.status)
+        self.assertEqual(bdata, res2.body)
+
+    def test_copy_object_invalid_body(self):
+        json_data = 'test'
+        res = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2/kasteel',
+                               json_data, headers={'content-type': 'application/json'}, expect_errors=True)
+        self.assertEqual(res.status, '400 Bad Request')
+        json_data = '{"path": "test"}'
+        res3 = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2/kasteel',
+                                json_data, headers={'content-type': 'application/json'}, expect_errors=True)
+        self.assertEqual(res3.status, '400 Bad Request')
+        json_data = json.dumps({
+            'host_url': 'bad host url',
+            'collection_key': 'TEST_COLLECTION',
+            'container_key': 'TEST_CONTAINER_ID',
+            'object_key': '200x300'
+        })
+        res4 = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2/kasteel',
+                                json_data, headers={'content-type': 'application/json'}, expect_errors=True)
+        self.assertEqual(res4.status, '400 Bad Request')
+
+    def test_copy_object_invalid_body_not_found(self):
+        json_data = json.dumps({
+            'host_url': 'http://localhost',
+            'collection_key': 'TEST_COLLECTION_INVALID',
+            'container_key': 'TEST_CONTAINER_ID',
+            'object_key': '200x300'
+        })
+        res = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2/kasteel',
+                               json_data, headers={'content-type': 'application/json'}, expect_errors=True)
+        self.assertEqual(res.status, '400 Bad Request')
+        self.assertIn('Collection TEST_COLLECTION_INVALID was not found', res.text)
+        json_data = json.dumps({
+            'host_url': 'http://localhost',
+            'collection_key': 'TEST_COLLECTION',
+            'container_key': 'TEST_CONTAINER_ID',
+            'object_key': '200'
+        })
+        res2 = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID2/kasteel',
+                                json_data, headers={'content-type': 'application/json'}, expect_errors=True)
+        self.assertEqual(res2.status, '400 Bad Request')
+        self.assertIn(
+            'Container - object (TEST_CONTAINER_ID - 200) combination was not found in Collection TEST_COLLECTION',
+            res2.text
+        )
+
     def test_delete_object(self):
         # create container and add object
         cres = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID')
@@ -204,4 +308,19 @@ class FunctionalTests(unittest.TestCase):
         if isinstance(container_key, unicode):
             container_key = str(container_key)
         self.assertTrue(uuid4hex.match(container_key))
+        print(res.json_body['uri'])
+
+    def test_create_object_with_id(self):
+        cres = self.testapp.put('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID')
+        self.assertEqual('200 OK', cres.status)
+        testdata = os.path.join(here, '../', 'fixtures/kasteel.jpg')
+        with open(testdata, 'rb') as f:
+            bdata = f.read()
+        res = self.testapp.post('/collections/TEST_COLLECTION/containers/TEST_CONTAINER_ID', bdata)
+        self.assertEqual('201 Created', res.status)
+        uuid4hex = re.compile('[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\Z', re.I)
+        object_key = res.json_body['object_key']
+        if isinstance(object_key, unicode):
+            object_key = str(object_key)
+        self.assertTrue(uuid4hex.match(object_key))
         print(res.json_body['uri'])