+++ tests/test_stores.py
@@ -52,6 +52,20 @@ class TestPairTreeStore(unittest.TestCase):
         object_list = self.store.list_object_keys_for_container(container_key)
         self.assertEqual(0, len(object_list))
 
+    def test_get_file_info(self):
+        here = os.path.dirname(__file__)
+        testdata = os.path.join(here, '../', 'fixtures/kasteel.jpg')
+        with open(testdata, 'rb') as f:
+            bdata = f.read()
+        container_key = 'testing'
+        object_key = 'metadata'
+        self.store.create_container(container_key)
+        self.store.create_object(container_key, object_key, bdata)
+        object_info = self.store.get_object_info(container_key, object_key)
+        self.assertEqual(11370, object_info['size'])
+        self.assertEqual('image/jpeg', object_info['mime'])
+        self.assertIn('time_last_modification', object_info)
+
     def test_update_scenario(self):
         container_key = 'testing'
         object_key = 'metadata'
@@ -108,6 +122,8 @@ class TestCephStore(unittest.TestCase):
         self.store.create_container(container_key)
         self.store.create_object(container_key, object_key, 'some test data')
         object_value = self.store.get_object(container_key, object_key)
+        self.assertEqual(None, object_value)
+        object_value = self.store.get_object_info(container_key, object_key)
         self.assertEqual(None, object_value)
         self.store.delete_object(container_key, object_key)
         object_list = self.store.list_object_keys_for_container(container_key)